Split conflicts code into a new file
